Add game version and save name config values.
Fixes issue where creating a new instance of the factorio server
will create a new save game(generating a new map) and has the possibility
of syncing that to the S3 bucket. Part of the EC2 instance install script
will now sync from S3 to the instance before attempting to create a new
game.

- Addd game version as config value (defaults to latest)
- Add game save name config value
- Update install script to sync existing save games to new instance
  before attempting to create new save game.